Remove internal deps on default ghc and stackage (#738)
Changes to the interface of haskell.nix (from the changelog.md file):

* Removed `sources.nixpkgs-default`, use `sources.nixpkgs` instead.
* Removed `./nixpkgs` directory, use  `(import ./. {}).sources`
  or `./nix/sources.nix` instead.
* Removes V1 interface for details on how to fix old code see:
    https://github.com/input-output-hk/haskell.nix/issues/709
* Removed defaultCompilerNixName.
* cabalProject, cabalProject', hackage-project and hackage-package
  now require a `compiler-nix-name` argument.
* `haskell-nix.tool` and `.tools` now require a `compiler-nix-name` argument.
  New functions `p.tool` and `p.tools` (where p is a project) do not.
  Like `shellFor { tools = ... }` they will use the compiler nix name
  from the project (including stack projects where it is derived from
  the resolver).
* `haskell-nix.alex` and `haskell-nix.happy` have been removed. Use
  `p.tool "alex" "3.2.5"` or `shellFor { tools = { alex = "3.2.5"; } }`.
* `haskell-nix.nix-tools` -> `haskell-nix.nix-tools.ghc883` (it includes
  the hpack exe now).
* `haskell-nix.cabal-install` -> 
  `p.tool "cabal" "3.2.0.0"` or `shellFor { tools = { cabal = "3.2.0.0"; } }`
* `haskell-nix.haskellNixRoots` -> `haskell-nix.roots ghc883` or `p.roots`

Other changes:

Adds hpack executable to the nix-tools derivations.

Adds a `cabal-hpack` test to make sure `hpack` works with
`cabalProject`.

Reduces the number of calls to `cabalProject` (particularly when
checking materialization), by giving internal tools a per-compiler
attribute.

Uses happy 1.19.12 when building newer ghc versions.

Updates cabal-install 3.2.0.0 to use the source from github that
is compatible with ghc 8.10.1.

Updates the docs for callCabalProjectToNix.

Adds a license mapping to fix a common warning.

+++ test/tests.sh
@@ -7,6 +7,13 @@ NIX_BUILD_ARGS="${NIX_BUILD_ARGS:-}"
 
 cd $(dirname $0)
 
+if [ "$#" != "1" ]; then
+  echo "Please pass a compiler-nix-name to use.  For example: ./test/test.sh ghc883"
+  exit 1
+fi
+
+GHC=$1
+
 printf "*** Cleaning package build directories..." >& 2
 rm -rvf */cabal.project.local */.ghc.environment* */dist */dist-newstyle */.stack-work
 echo >& 2
@@ -16,11 +23,12 @@ nix build $NIX_BUILD_ARGS \
    -I . -I .. \
    --option restrict-eval true \
    --option allowed-uris "https://github.com/NixOS https://github.com/input-output-hk" \
-   --no-link --keep-going -f default.nix
+   --no-link --keep-going -f default.nix \
+   --argstr compiler-nix-name $GHC
 echo >& 2
 
 printf "*** Running the unit tests... " >& 2
-res=$(nix-instantiate --eval --json --strict ./default.nix -A unit.tests)
+res=$(nix-instantiate --eval --json --strict ./default.nix --argstr compiler-nix-name $GHC -A unit.tests)
 num_failed=$(jq length <<< "$res")
 if [ $num_failed -eq 0 ]; then
   printf "PASSED\n" >& 2
@@ -33,6 +41,7 @@ fi
 printf "*** Checking that a nix-shell works for runghc...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A with-packages.test-shell \
     --run 'runghc with-packages/Point.hs'
 echo >& 2
@@ -40,6 +49,7 @@ echo >& 2
 printf "*** Checking that a nix-shell works for cabal...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A with-packages.test-shell \
     --run 'echo CABAL_CONFIG=$CABAL_CONFIG && type -p ghc && cd with-packages && cabal new-build'
 echo >& 2
@@ -47,6 +57,7 @@ echo >& 2
 printf "*** Checking that a nix-shell works for cabal (doExactConfig component)...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A with-packages.test-shell-dec \
     --run 'echo CABAL_CONFIG=$CABAL_CONFIG && echo GHC_ENVIRONMENT=$GHC_ENVIRONMENT && cd with-packages && cabal new-build'
 echo >& 2
@@ -55,6 +66,7 @@ printf "*** Checking that a nix-shell works for a project with test-suite build-
 printf "!!! This is expected to fail until https://github.com/input-output-hk/haskell.nix/issues/231 is resolved! \n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A cabal-22.shell \
     --run 'cd cabal-22 && cabal new-build all --enable-tests --enable-benchmarks' \
     || true
@@ -63,6 +75,7 @@ echo >& 2
 printf "*** Checking that a nix-shell works for a multi-target project...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A cabal-simple.test-shell \
     --run 'cd cabal-simple && cabal new-build'
 echo >& 2
@@ -70,6 +83,7 @@ echo >& 2
 printf "*** Checking shellFor works for a cabal project, multiple packages...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A shell-for.env \
     --run 'cd shell-for && cabal new-build all'
 echo >& 2
@@ -77,6 +91,7 @@ echo >& 2
 printf "*** Checking shellFor works for a cabal project, single package...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A shell-for.envPkga \
     --run 'cd shell-for && cabal new-build --project=single.project all'
 echo >& 2
@@ -84,14 +99,15 @@ echo >& 2
 printf "*** Checking shellFor has a working hoogle index...\n" >& 2
 nix-shell $NIX_BUILD_ARGS \
     --pure ./default.nix \
+    --argstr compiler-nix-name $GHC \
     -A shell-for.env \
     --run 'hoogle ConduitT | grep Data.Conduit'
 echo >& 2
 
 printf "*** Checking shellFor does not depend on given packages...\n" >& 2
-drva=$(nix-instantiate ./default.nix -A shell-for.env)
+drva=$(nix-instantiate ./default.nix --argstr compiler-nix-name $GHC -A shell-for.env)
 echo "-- hello" >> shell-for/pkga/PkgA.hs
-drvb=$(nix-instantiate ./default.nix -A shell-for.env)
+drvb=$(nix-instantiate ./default.nix --argstr compiler-nix-name $GHC -A shell-for.env)
 sed -i -e '/-- hello/d' shell-for/pkga/PkgA.hs
 if [ "$drva" != "$drvb" ]; then
     printf "FAIL\nShell derivations\n$drva\n$drvb\n are not identical.\n" >& 2
@@ -101,19 +117,19 @@ else
 fi
 
 printf "*** Checking the maintainer scripts...\n" >& 2
-nix build $NIX_BUILD_ARGS --no-link --keep-going -f ../build.nix maintainer-scripts
+nix build $NIX_BUILD_ARGS --no-link --keep-going -f ../build.nix --argstr compiler-nix-name $GHC maintainer-scripts
 echo >& 2
 
 printf "*** Checking that plan construction works with extra Hackages...\n" >& 2
 nix build $NIX_BUILD_ARGS --no-link \
-    -f ./default.nix \
-    extra-hackage.plan-nix
+    -f ./default.nix --argstr compiler-nix-name $GHC \
+    extra-hackage.run.project.plan-nix
 echo >& 2
 
 printf "*** Checking that package with extra Hackages can be build...\n" >& 2
 nix build $NIX_BUILD_ARGS --no-link \
-    -f ./default.nix \
-    extra-hackage.hsPkgs.external-package-user.components.all
+    -f ./default.nix --argstr compiler-nix-name $GHC \
+    extra-hackage.run.project.hsPkgs.external-package-user.components.all
 echo >& 2
 
 printf "\n*** Finished successfully\n" >& 2
